학원 코드.
Camera를 3D에 맞게 수정.
Tool Camera 기능 구현.

+++ DirectX11/03. Tool/BtnView.cpp
@@ -64,12 +64,12 @@ void CBtnView::DoDataExchange(CDataExchange* pDX)
 	CRect buttonAddLayerRect;
 	buttonAddLayer.GetWindowRect(buttonAddLayerRect);
 
-	float fAccWidth = rt1.Width() + rt2.Width() + rt3.Width();
+	int fAccWidth = rt1.Width() + rt2.Width() + rt3.Width();
 
 	CRect rtBtnView;
 	GetClientRect(rtBtnView);
 
-	float fStartX = (rtBtnView.Width() / 2.f) - (fAccWidth / 2.f);
+	int fStartX = (rtBtnView.Width() / 2) - (fAccWidth / 2);
 
 	m_btnPlay.SetWindowPos(nullptr, fStartX, 10, rt1.Width(), rt1.Height(), 0);
